+++ python/ipywidgets/ipywidgets/widgets/__init__.py
@@ -19,7 +19,7 @@ from .widget_datetime import DatetimePicker
 from .widget_time import TimePicker
 from .widget_output import Output
 from .widget_selection import RadioButtons, ToggleButtons, ToggleButtonsStyle, Dropdown, Select, SelectionSlider, SelectMultiple, SelectionRangeSlider
-from .widget_selectioncontainer import Tab, Accordion, Stacked
+from .widget_selectioncontainer import Tab, Accordion, Stack
 from .widget_string import HTML, HTMLMath, Label, Text, Textarea, Password, Combobox
 from .widget_controller import Controller
 from .interaction import interact, interactive, fixed, interact_manual, interactive_output